futher code splitting, various bug fixes
